web3js: Rewrite accounts
