+++ web3js-ext/src/types.ts
@@ -16,7 +16,9 @@ import {
   sign,
   signTransaction,
   TxData,
-  TypedTransaction,
+  Transaction as LegacyTransaction,
+  AccessListEIP2930Transaction,
+  FeeMarketEIP1559Transaction,
   Wallet,
   Web3Account,
 } from "web3-eth-accounts";
@@ -27,7 +29,7 @@ import { Personal } from "web3-eth-personal";
 import { Net } from "web3-net";
 import { KeyStore, Bytes, Transaction } from "web3-types";
 
-import { KlaytnTx } from "./klaytn_tx";
+import { KlaytnTypedTransaction } from "./accounts/klaytn_tx";
 
 // Type analogous to web3-eth-accounts/src/types.ts:Web3Account
 // Designed for the "account object" returned by
@@ -148,6 +150,8 @@ export interface KlaytnTxData extends TxData {
 
 // The child classes of BaseTransaction.
 // Used inside signTransaction() and signTransactionAsFeePayer()
-export type KlaytnTypedTransaction =
-	| TypedTransaction // Type 0 (LegacyTransaction), Type 1, Type 2
-	| KlaytnTx; // Klaytn TxTypes
+export type TypedTransaction =
+  | LegacyTransaction // type 0
+  | AccessListEIP2930Transaction // type 1
+  | FeeMarketEIP1559Transaction // type 2
+  | KlaytnTypedTransaction; // Klaytn TxTypes
