+++ web3js-ext/test/account.spec.ts
@@ -22,7 +22,7 @@ const gasPrice = 25e9;
 const gasLimit = 50000;
 const chainId = 1001;
 const senderTxHashRLP = "0x09f87a8204d219830f4240947b65b75d204abed71587c9e519a89277766ee1d00a94a94f5374fce5edbc8e2a8697c15331677e6ebf0bf845f84325a09f8e49e2ad84b0732984398749956e807e4b526c786af3c5f7416b293e638956a06bf88342092f6ff9fabe31739b2ebfa1409707ce54a54693e91a6b9bb77df0e7";
-const txSignatures = [["0x25", "0x9f8e49e2ad84b0732984398749956e807e4b526c786af3c5f7416b293e638956", "0x6bf88342092f6ff9fabe31739b2ebfa1409707ce54a54693e91a6b9bb77df0e7"]];
+const txSignatures = [["0x7f5", "0x9f8e49e2ad84b0732984398749956e807e4b526c786af3c5f7416b293e638956", "0x6bf88342092f6ff9fabe31739b2ebfa1409707ce54a54693e91a6b9bb77df0e7"]];
 const tx0 = { type: 0, from, to, value, gasPrice };
 const tx9 = { type: 9, from, to, value, gasPrice, txSignatures };
 const lightKdf: CipherOptions = { kdf: "scrypt", n: 2, p: 1 }; // For faster testing
